requirements.txt
futures==3.0.5
tornado==4.3
kazoo==2.2.1
kafka-python==1.3.1
redis==2.10.5